Hist_Segmentation
This script is designed to create histograms displaying the frequency distribution of pixel values in an image. The image is divided into a 3x3 grid, and a separate histogram is calculated for each block to obtain more accurate results.
To use this script, simply download the Hist_Segmentation.py file and place it in the same directory as the images you wish to analyze (named model.png and test.png). Then, run the script and the histograms will be displayed.
Run code in terminal : python Hist_Segmentation.py